refactored application files from com.auth0 to com.goat and renamed package with name convention
