+++ ThirstyGoat/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.auth0.ThirstyGoat">
+    package="com.goat.thirsty_goat">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
@@ -14,7 +14,7 @@
         android:theme="@style/AppTheme"
         tools:ignore="AllowBackup">
         <activity
-            android:name="com.auth0.ThirstyGoat.StartActivity"
+            android:name="com.goat.thirsty_goat.StartActivity"
             android:noHistory="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -22,8 +22,8 @@
             </intent-filter>
         </activity>
 
-        <activity android:name="com.auth0.ThirstyGoat.LoginActivity" />
-        <activity android:name="com.auth0.ThirstyGoat.MainActivity" />
+        <activity android:name="com.goat.thirsty_goat.LoginActivity" />
+        <activity android:name="com.goat.thirsty_goat.MainActivity" />
 
         <!-- Auth0 Lock -->
         <activity
@@ -40,7 +40,7 @@
 
                 <data
                     android:host="@string/auth0_domain"
-                    android:pathPrefix="/android/com.auth0.ThirstyGoat/callback"
+                    android:pathPrefix="/android/com.goat.thirsty_goat/callback"
                     android:scheme="https" />
             </intent-filter>
         </activity>
